titan-poss-poss-web-f78a2c9ef984/titan-poss-poss-web-f78a2c9ef984/libs/eposs/amendment-config/data-access-amendment-config/src/lib/+state/amendment-config.selectors.ts
import { createSelector } from '@ngrx/store';
import { selectAmendmentConfigState } from './amendment-config.reducer';

const selectAmendmentConfiguration = createSelector(
  selectAmendmentConfigState,
  state => state.amendmentConfigValue
);
const selectError = createSelector(
  selectAmendmentConfigState,
  state => state.error
);
const selectIsLoading = createSelector(
  selectAmendmentConfigState,
  state => state.isLoading
);
const selectHasUpdated = createSelector(
  selectAmendmentConfigState,
  state => state.hasUpdated
);
export const AmendmentConfigurationSelectors = {
  selectAmendmentConfiguration,
  selectIsLoading,
  selectHasUpdated,
  selectError
};
